pyproject.toml
[tool.poetry]
name = "test-task-telegram-planfix"
version = "0.1.0"
description = ""
authors = ["Victor Polyakov <polyakov.victor21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = {extras = ["async"], version = "^3.0.0"}
telethon = "^1.33.1"
python-dotenv = "^1.0.0"
async = "^0.6.2"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
